feat: classe de error

+++ src/storage/group/groupCreate.ts
@@ -1,16 +1,23 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { AppError } from '@utils/AppError';
 
 import { GROUP_COLLECTION } from '@storage/storageConfig';
 import { groupsGetAll } from './groupsGetAll';
 
-export async function groupCreate(newGroup: string){
+export async function groupCreate(newGroup: string) {
   try {
-     const storedGroups = await groupsGetAll();
+    const storedGroups = await groupsGetAll();
+
+    const groupAlredyExists = storedGroups.includes(newGroup);
+
+    if (groupAlredyExists){
+      throw new AppError('Já existe um grupo cadastrado com esse nome.') ;
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup]);
 
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
-  } catch(error){
+  } catch (error) {
     throw error;
   }
 }
